contoso_agents/network_performance_agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rogers-network-performance-agent"
version = "0.1.0"
description = "Rogers Network Performance Agent - Comprehensive network diagnostics for internet troubleshooting"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.2.6",
    "azure-ai-agents>=1.1.0b2",
    "azure-ai-projects>=1.0.0b12",
    "azure-identity>=1.19.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.41.0",
    "uvicorn>=0.34.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
